third_party/blink/web_tests/external/wpt/service-workers/service-worker/xhr-response-url.https-expected.txt

This is a testharness.js-based test.
FAIL XHR responseURL should be the response URL assert_equals: expected "https://web-platform.test:8444/service-workers/service-worker/resources/dummy.txt" but got "https://web-platform.test:8444/service-workers/service-worker/resources/test?url=https%3A%2F%2Fweb-platform.test%3A8444%2Fservice-workers%2Fservice-worker%2Fresources%2Fdummy.txt"
Harness: the test ran to completion.
